apps/cic-cache/requirements.txt
alembic==1.4.2
confini~=0.5.3
uwsgi==2.0.19.1
moolb~=0.2.0
cic-eth-registry~=0.6.6
SQLAlchemy==1.3.20
semver==2.13.0
psycopg2==2.8.6
celery==4.4.7
redis==3.5.3
chainsyncer[sql]~=0.0.7
erc20-faucet~=0.3.2
chainlib-eth~=0.0.15
eth-address-index~=0.2.4
okota~=0.2.5
